# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "balldontlie API",
        "description": "Teams, Players and Games management",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/teams": {
            "get": {
                "tags": [
                    "Teams"
                ],
                "summary": "Get a list of teams",
                "operationId": "SearchTeams",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Invalid request"
                    }
                }
            },
            "post": {
                "tags": [
                    "Teams"
                ],
                "summary": "Create one team",
                "operationId": "CreateTeam",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "": {
                                        "properties": {
                                            "conference": {
                                                "type": "string"
                                            },
                                            "division": {
                                                "type": "string"
                                            },
                                            "city": {
                                                "type": "string"
                                            },
                                            "name": {
                                                "type": "string"
                                            },
                                            "full_name": {
                                                "type": "string"
                                            },
                                            "abbreviation": {
                                                "type": "string"
                                            }
                                        },
                                        "type": "object"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "conference": "West",
                                    "division": "Southwest",
                                    "city": "Los Angeles",
                                    "name": "Raptors",
                                    "full_name": "Portland Trail Blazers",
                                    "abbreviation": "PTB"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Invalid request"
                    }
                }
            }
        },
        "/api/v1/teams/{teamId}": {
            "get": {
                "tags": [
                    "Teams"
                ],
                "summary": "Get a one team",
                "operationId": "GetTeamDetails",
                "parameters": [
                    {
                        "name": "teamId",
                        "in": "path",
                        "description": "Team ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        },
                        "example": "1"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Invalid request"
                    }
                }
            },
            "put": {
                "tags": [
                    "Teams"
                ],
                "summary": "Update one team",
                "operationId": "UpdateTeamDetails",
                "parameters": [
                    {
                        "name": "teamId",
                        "in": "path",
                        "description": "Team ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        },
                        "example": "1"
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "": {
                                        "properties": {
                                            "conference": {
                                                "type": "string"
                                            },
                                            "division": {
                                                "type": "string"
                                            },
                                            "city": {
                                                "type": "string"
                                            },
                                            "name": {
                                                "type": "string"
                                            },
                                            "full_name": {
                                                "type": "string"
                                            },
                                            "abbreviation": {
                                                "type": "string"
                                            }
                                        },
                                        "type": "object"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "conference": "Mid-East",
                                    "division": "Northwest",
                                    "city": "Los Angeles",
                                    "name": "Raptors",
                                    "full_name": "Portland Trail Blazers",
                                    "abbreviation": "PTB"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Invalid request"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Teams"
                ],
                "summary": "Delete a one team",
                "operationId": "DeleteTeamDetails",
                "parameters": [
                    {
                        "name": "teamId",
                        "in": "path",
                        "description": "Team ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        },
                        "example": "1"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Invalid request"
                    }
                }
            }
        }
    },
    "components": {
        "securitySchemes": {
            "token": {
                "type": "apiKey",
                "name": "Authorization",
                "in": "header"
            }
        }
    }
}